working pipeline ready for run
