+++ data_processing/infraxclimate/nasa_nex/constants.py
@@ -2,3 +2,6 @@ X_DIM = "lon"
 Y_DIM = "lat"
 TIME_DIM = "time"
 METADATA_KEY = "UW_CRL_DERIVED"
+
+# -999 used for 'historical' scenario (1950-2014)
+SSPS = [-999, 126, 245, 370, 585]
